Avoid travis to install npm deps twice

+++ tests/tests.bash
@@ -3,11 +3,13 @@
 # Go back to root project directory
 cd `dirname $0`/../
 
-echo "Removing node modules..."
-rm -rf node_modules/
+if [[ '--reinstall' == *"$*"* ]]; then
+    echo "Removing node modules..."
+    rm -rf node_modules/
 
-echo "Installing dependencies..."
-npm install
+    echo "Installing dependencies..."
+    npm install
+fi
 
 echo "Preparing tests..."
 node tests/prepare_tests.js
